refactor(message review): refactor components into sub folder
